+++ src/LDL/Framework/Base/Contracts/JsonFileFactoryInterface.php
@@ -1,0 +1,15 @@
+<?php declare(strict_types=1);
+
+namespace LDL\Framework\Base\Contracts;
+
+use LDL\Framework\Base\Exception\JsonFileFactoryException;
+
+interface JsonFileFactoryInterface
+{
+    /**
+     * @param string $file
+     * @throws JsonFileFactoryException
+     * @return JsonFileFactoryInterface
+     */
+    public static function fromJsonFile(string $file) : JsonFileFactoryInterface;
+}